advised refactoring
