+++ src/main/java/gov/nih/nci/evs/api/service/StardogElasticLoadServiceImpl.java
@@ -12,22 +12,16 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.DefaultParser;
-import org.apache.commons.cli.HelpFormatter;
-import org.apache.commons.cli.Options;
-import org.apache.commons.cli.ParseException;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.SpringApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import gov.nih.nci.evs.api.Application;
 import gov.nih.nci.evs.api.model.Concept;
 import gov.nih.nci.evs.api.model.ConceptMinimal;
 import gov.nih.nci.evs.api.model.IncludeParam;
@@ -44,10 +38,10 @@ import gov.nih.nci.evs.api.util.TerminologyUtils;
  * @author Arun
  */
 @Service
-public class ElasticLoadServiceImpl extends BaseLoaderService implements ElasticLoadService {
+public class StardogElasticLoadServiceImpl extends BaseLoaderService {
 
   /** the logger *. */
-  private static final Logger logger = LoggerFactory.getLogger(ElasticLoadServiceImpl.class);
+  private static final Logger logger = LoggerFactory.getLogger(StardogElasticLoadServiceImpl.class);
 
   /** the concepts download location *. */
   @Value("${nci.evs.bulkload.conceptsDir}")
@@ -304,78 +298,6 @@ public class ElasticLoadServiceImpl extends BaseLoaderService implements Elastic
   }
 
   /**
-   * Clean stale indexes.
-   *
-   * @throws Exception the exception
-   */
-  public void cleanStaleIndexes() throws Exception {
-    List<IndexMetadata> iMetas = null;
-    iMetas = termUtils.getStaleTerminologies();
-
-    if (CollectionUtils.isEmpty(iMetas))
-      return;
-
-    logger.info("Removing stale terminologies: " + iMetas);
-
-    for (IndexMetadata iMeta : iMetas) {
-      String indexName = iMeta.getIndexName();
-      String objectIndexName = iMeta.getObjectIndexName();
-
-      // objectIndexName will be NULL if terminology object is not part of
-      // IndexMetadata
-      // temporarily required to accommodate change in IndexMetadata object
-      if (objectIndexName == null) {
-        objectIndexName = "evs_object_" + indexName.replace("concept_", "");
-      }
-
-      // delete objects index
-      boolean result = operationsService.deleteIndex(objectIndexName);
-
-      if (!result) {
-        logger.warn("Deleting objects index {} failed!", objectIndexName);
-        continue;
-      }
-
-      // delete concepts index
-      result = operationsService.deleteIndex(indexName);
-
-      if (!result) {
-        logger.warn("Deleting concepts index {} failed!", indexName);
-        continue;
-      }
-
-      // delete metadata object
-      esQueryService.deleteIndexMetadata(indexName);
-    }
-  }
-
-  /**
-   * Update latest flag.
-   *
-   * @throws Exception the exception
-   */
-  public void updateLatestFlag() throws Exception {
-    // update latest flag
-    logger.info("Updating latest flags on all metadata objects");
-    List<IndexMetadata> iMetas = esQueryService.getIndexMetadata(true);
-
-    if (CollectionUtils.isEmpty(iMetas))
-      return;
-
-    Terminology latest = termUtils.getLatestTerminology(false);
-    for (IndexMetadata iMeta : iMetas) {
-      if (iMeta.getTerminology() != null) {
-        iMeta.getTerminology()
-            .setLatest(iMeta.getTerminology().getIndexName().equals(latest.getIndexName()));
-      }
-    }
-
-    operationsService.bulkIndex(iMetas, ElasticOperationsService.METADATA_INDEX,
-        ElasticOperationsService.METADATA_TYPE, IndexMetadata.class);
-
-  }
-
-  /**
    * Task to load a batch of concepts to elasticsearch.
    *
    * @author Arun
@@ -447,86 +369,8 @@ public class ElasticLoadServiceImpl extends BaseLoaderService implements Elastic
     }
   }
 
-  /**
-   * the main method to trigger elasticsearch load via command line *.
-   *
-   * @param args the command line arguments
-   */
-  public static void main(String[] args) {
-    Options options = prepareOptions();
-    CommandLine cmd;
-    try {
-      cmd = new DefaultParser().parse(options, args);
-    } catch (ParseException e) {
-      logger.error("{}; Try -h or --help to learn more about command line options available.",
-          e.getMessage());
-      return;
-    }
-
-    if (cmd.hasOption('h')) {
-      printHelp(options);
-      return;
-    }
-
-    ApplicationContext app = SpringApplication.run(Application.class, new String[0]);
-
-    try {
-      // get the bean by type
-      ElasticLoadServiceImpl loadService = app.getBean(ElasticLoadServiceImpl.class);
-
-      ElasticLoadConfig config = buildConfig(cmd, loadService.CONCEPTS_OUT_DIR);
-
-      if (StringUtils.isBlank(config.getTerminology())) {
-        logger.error(
-            "Terminology (-t or --terminology) is required! Try -h or --help to learn more about command line options available.");
-        return;
-      }
-
-      TerminologyUtils termUtils = app.getBean(TerminologyUtils.class);
-      Terminology term = termUtils.getTerminology(config.getTerminology(), false);
-      HierarchyUtils hierarchy = loadService.sparqlQueryManagerService.getHierarchyUtils(term);
-      loadService.loadConcepts(config, term, hierarchy);
-      loadService.loadObjects(config, term, hierarchy);
-      loadService.cleanStaleIndexes();
-      loadService.updateLatestFlag();
-    } catch (Exception e) {
-      logger.error(e.getMessage(), e);
-      throw new RuntimeException(e);
-    } finally {
-      SpringApplication.exit(app);
-    }
-  }
-
-  /**
-   * prepare command line options available.
-   *
-   * @return the options
-   */
-  private static Options prepareOptions() {
-    Options options = new Options();
-
-    options.addOption("f", "forceDeleteIndex", false,
-        "Force delete index if index already exists.");
-    options.addOption("h", "help", false, "Show this help information and exit.");
-    options.addOption("r", "realTime", false, "Keep for backwards compabitlity. No Effect.");
-    options.addOption("t", "terminology", true, "The terminology (ex: ncit_20.02d) to load.");
-
-    return options;
-  }
-
-  /**
-   * print command line help information.
-   *
-   * @param options the options available
-   */
-  private static void printHelp(Options options) {
-    HelpFormatter formatter = new HelpFormatter();
-    formatter.printHelp("java -jar $DIR/evsrestapi-*.jar", options);
-    return;
-  }
-
   @Override
-  void setUpConceptLoading(ApplicationContext app, CommandLine cmd) throws Exception {
+  public void setUpConceptLoading(ApplicationContext app, CommandLine cmd) throws Exception {
   try {
       // get the bean by type
